+++ CHANGELOG.md
@@ -8,6 +8,27 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+## [Release-51][release-51] (production-2025-09-18.6482)
+
+### Changed
+
+- Fix accessibility issue in the 'About the data we use' page.
+- Alternate row colours on tables and added table header colours.
+- For Ofsted pages changed "Joined before" tag from grey to orange.
+
+## [Release-50][release-50] (production-2025-09-10.6449)
+
+### Changed
+
+- Added governance turnover to trust pages.
+- Added additional information to the individual schools' ofsted downloads.
+
+## [Release-49][release-49] (production-2025-09-04.6416)
+
+### Changed
+
+- Changed the design of the Trust details, Trust summary and Reference numbers pages to use the gov uk summary list design
+
 ## [Release-48][release-48] (production-2025-08-28.6382)
 
 ### Changed
@@ -455,7 +476,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Updated text on "404 - Not Found" page to adhere to DfE design pattern and added the not found url to the support email template.
 
 [unreleased]:
-  https://github.com/DFE-Digital/find-information-about-academies-and-trusts/compare/production-2025-08-28.6382...HEAD
+  https://github.com/DFE-Digital/find-information-about-academies-and-trusts/compare/production-2025-09-18.6482...HEAD
 [release-1]:
   https://github.com/DFE-Digital/find-information-about-academies-and-trusts/releases/tag/production-2024-07-18.2517
 [release-2]:
@@ -552,3 +573,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   https://github.com/DFE-Digital/find-information-about-schools-and-trusts/releases/tag/production-2025-08-21.6341
 [release-48]:
   https://github.com/DFE-Digital/find-information-about-schools-and-trusts/releases/tag/production-2025-08-28.6382
+[release-49]:
+  https://github.com/DFE-Digital/find-information-about-schools-and-trusts/releases/tag/production-2025-09-04.6416
+[release-50]:
+  https://github.com/DFE-Digital/find-information-about-schools-and-trusts/releases/tag/production-2025-09-10.6449
+[release-51]:
+  https://github.com/DFE-Digital/find-information-about-schools-and-trusts/releases/tag/production-2025-09-18.6482